Update main.cpp
more bug fixes, but what is not working properly yet is the filepath update for the micro overlay.  if you have the name set as 0-Status-Monitor-Overlay.ovl for instance, everything works, and it shows the micro overlay.  the micro overlay works as well and you can return back to status monitor.  however trying to exit after that will crash.  Using the default name Status-Monitor-Overlay.ovl works still.  There is one step in the chain that is not reading the updates to the filepath when returning back to the menu from the micro overlay